@@ -1,75 +1,77 @@
+using Brainf.Memory;
 using Xunit;
 
-namespace Brainf.Tests
+namespace Brainf.Tests;
+
+public sealed class BrainfMemoryTests
 {
-    public sealed class BrainfMemoryTests
+    [Fact]
+    public void Positive_Pointer()
     {
-        [Fact]
-        public void Positive_Pointer()
+        // Arrange
+        var memory = new BrainfMemory();
+        var sum = 0;
+
+        // Act
+        for (var i = 0; i < 10; i++)
         {
-            // Arrange
-            var memory = new BrainfMemory();
-            var sum = 0;
-
-            // Act
-            for (var i = 0; i < 10; i++)
-            {
-                memory.CellValue = i;
-                memory.Pointer = i;
-            }
-
-            for (var i = 9; i >= 0; i--)
-            {
-                memory.Pointer = i;
-                sum += memory.CellValue;
-            }
-
-            // Assert
-            Assert.Equal(45, sum);
+            memory.Pointer = i;
+            memory.CellValue = i;
         }
+
+        for (var i = 9; i >= 0; i--)
+        {
+            memory.Pointer = i;
+            sum += memory.CellValue;
+        }
+
+        // Assert
+        Assert.Equal(45, sum);
+        Assert.Equal(0, memory.Pointer);
+        Assert.Equal(0, memory.CellValue);
+    }
         
-        [Fact]
-        public void Negative_Pointer()
+    [Fact]
+    public void Negative_Pointer()
+    {
+        // Arrange
+        var memory = new BrainfMemory();
+        var sum = 0;
+
+        // Act
+        for (var i = -1; i >= -10; i--)
         {
-            // Arrange
-            var memory = new BrainfMemory();
-            var sum = 0;
-
-            // Act
-            for (var i = -1; i >= -10; i--)
-            {
-                memory.CellValue = -(i + 1);
-                memory.Pointer = i;
-            }
-
-            for (var i = -10; i <= -1 ; i++)
-            {
-                memory.Pointer = i;
-                sum += memory.CellValue;
-            }
-
-            // Assert
-            Assert.Equal(45, sum);
+            memory.CellValue = -(i + 1);
+            memory.Pointer = i;
         }
 
-        [Fact]
-        public void Capacity()
+        for (var i = -10; i <= -1 ; i++)
         {
-            // Arrange
-            var memory = new BrainfMemory(10);
-
-            // Act
-            var capacity = memory.Capacity;
-
-            for (var i = 0; i < 11; i++)
-            {
-                memory.CellValue = i;
-                memory.Pointer++;
-            }
-
-            // Assert
-            Assert.Equal(10, capacity);
-            Assert.Equal(20, memory.Capacity);
+            memory.Pointer = i;
+            sum += memory.CellValue;
         }
+
+        // Assert
+        Assert.Equal(45, sum);
+    }
+
+    [Fact]
+    public void Capacity()
+    {
+        // Arrange
+        var memory = new BrainfMemory(10);
+
+        // Act
+        var capacity = memory.Capacity;
+
+        for (var i = 0; i < 11; i++)
+        {
+            memory.CellValue = i;
+            memory.Pointer++;
+        }
+
+        // Assert
+        Assert.Equal(10, capacity);
+        Assert.Equal(20, memory.Capacity);
     }
 }
